BREAKING CHANGE: Support for flexible server
This commit switches from the deprecated single server solution to the new flexible server.

The configuration is nearly the same as before, but the different network handling makes the configuration incompatible.

Warning: Using this version DESTROYS the previously deployed server, so a manual migration using dumps and restore is required.

Additionally, additional mysql configurations are supported in this as well.